+++ packages/nesp_sdk_dart_core/lib/nesp_sdk_dart_core.dart
@@ -19,6 +19,9 @@ export 'src/exceptions.dart';
 export 'src/result.dart';
 export 'src/stream_subscription_registry.dart';
 export 'src/initializer.dart';
+export 'src/debugger.dart';
+export 'src/daemon_service.dart';
+export 'src/daemon_task.dart';
 
 export 'lang.dart';
 export 'data.dart';